requirements.txt
requests==2.32.3
dashscope==1.25.8
Pillow==11.1.0
colorlog==6.9.0
schedule==1.2.2
